feat(stylus2scss): complete flow with improved options

+++ libs/tools/stylus2scss/src/converter/stylus2scss.ts
@@ -1,16 +1,21 @@
 import { readFile, writeFile } from '@myndpm/utils';
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { concatMap, map } from 'rxjs/operators';
 import { Parser } from 'stylus';
 import { findMixin, findVariable } from '../parser/utils';
-import { Options } from '../schema';
+import { Schema } from '../schema';
+import { logInfo } from '../utils';
 import { converter } from './converter';
 
 const GLOBAL_MIXINS: string[] = []
 const GLOBAL_VARIABLES: string[] = []
 
-export function stylusParse(file: string): Observable<any> {
-  return readFile(file).pipe(
+export function stylusParse(opts: Required<Schema>): Observable<any> {
+  if (opts.dryRun) {
+    logInfo('> Parsing stylus');
+    return of(true);
+  }
+  return readFile(opts.file).pipe(
     map((data) => {
       const ast = new Parser(data).parse();
       ast.nodes.forEach(node => {
@@ -21,12 +26,16 @@ export function stylusParse(file: string): Observable<any> {
   );
 }
 
-export function stylusConvert(options: Required<Options>): Observable<any> {
-  return readFile(options.file).pipe(
+export function stylusConvert(opts: Required<Schema>): Observable<any> {
+  if (opts.dryRun) {
+    logInfo('> Converting stylus');
+    return of(true);
+  }
+  return readFile(opts.file).pipe(
     concatMap((content) => {
       return writeFile(
-        options.file,
-        converter(content, options, GLOBAL_VARIABLES, GLOBAL_MIXINS),
+        opts.file,
+        converter(content, opts, GLOBAL_VARIABLES, GLOBAL_MIXINS),
       );
     }),
   );
